# Remove logging to the console, use pooled connections in 'nodejs-postgres --type update'